递归处理队列 (#64)
* document mutation observer

* fix type

* 使用递归处理队列

* cleanup

* cleanup

* fix type
